+++ db/sqlc/querier.go
@@ -17,12 +17,12 @@ type Querier interface {
 	CreateSession(ctx context.Context, arg CreateSessionParams) (Session, error)
 	CreateTransfer(ctx context.Context, arg CreateTransferParams) (Transfer, error)
 	CreateUser(ctx context.Context, arg CreateUserParams) (User, error)
-	DeleteAccount(ctx context.Context, id int64) error
-	DeleteUser(ctx context.Context, id int64) error
-	GetAccount(ctx context.Context, id int64) (Account, error)
-	GetEntry(ctx context.Context, id int64) (Entry, error)
+	DeleteAccount(ctx context.Context, id int32) error
+	DeleteUser(ctx context.Context, id int32) error
+	GetAccount(ctx context.Context, id int32) (Account, error)
+	GetEntry(ctx context.Context, id int32) (Entry, error)
 	GetSession(ctx context.Context, id uuid.UUID) (Session, error)
-	GetTransfer(ctx context.Context, id int64) (Transfer, error)
+	GetTransfer(ctx context.Context, id int32) (Transfer, error)
 	GetUser(ctx context.Context, username string) (User, error)
 	ListAccount(ctx context.Context, arg ListAccountParams) ([]Account, error)
 	ListEntries(ctx context.Context, arg ListEntriesParams) ([]Entry, error)
