+++ db/sqlc/sessions.sql.go
@@ -28,7 +28,7 @@ INSERT INTO sessions (
 
 type CreateSessionParams struct {
 	ID           uuid.UUID `json:"id"`
-	UserID       int64     `json:"user_id"`
+	UserID       int32     `json:"user_id"`
 	RefreshToken string    `json:"refresh_token"`
 	UserAgent    string    `json:"user_agent"`
 	ClientIp     string    `json:"client_ip"`
